+++ changelog.md
@@ -2,6 +2,8 @@
 
 Version | Changes
 ---|---
+15.1.0 | `ref` + `channel` merged into `bus` - accessor/channel primitive. `symbol.bus` and `symbol.observable` introduced.
+15.0.0 | Reactive `html` effect, `input`, better more optimized core.
 14.0.0 | $ is considered a selector observer special effect, added the generic effect `fx` and observables `state`, `calc`, `ref`, `store` etc.
 13.0.0 | Orphaned minimal selector-observer implementation, single $ entry.
 12.0.0 | Internalized hooks via augmentor, dropped unihooks dep. Equiv. of [hooked-elements](https://github.com/WebReflection/hooked-elements).